+++ app/Http/Controllers/Admin/ProfilesController.php
@@ -5,7 +5,6 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\UpdateProfileRequest;
 use Illuminate\Support\Facades\Storage;
-use Illuminate\Http\Request;
 use Intervention\Image\ImageManagerStatic as Image;
 use App\User;
 
@@ -18,6 +17,8 @@ class ProfilesController extends Controller
 
     public function index()
     {
+        abort_unless(\Gate::allows('profile_access'), 403);
+
         $user = auth()->user();
 
         return view('admin.profiles.index', compact('user'));
@@ -25,6 +26,8 @@ class ProfilesController extends Controller
 
     public function edit()
     {
+        abort_unless(\Gate::allows('profile_edit'), 403);
+
         $user = auth()->user();
 
         return view('admin.profiles.edit', compact('user'));
@@ -32,20 +35,24 @@ class ProfilesController extends Controller
 
     public function update(UpdateProfileRequest $request, User $user)
     {
+        abort_unless(\Gate::allows('profile_edit'), 403);
+
         $user = auth()->user();
 
         if (!empty($request->avatar)) {
-            $avatarName = $user->id.'_avatar'.time().'.'.request()->avatar->getClientOriginalExtension();
+            $avatarName = $user->id . '_avatar' . time() . '.' . request()->avatar->getClientOriginalExtension();
             if ($user->avatar != $avatarName && $user->avatar != 'user.png') {
-                Storage::delete('public/avatars/'.$user->avatar);
+                Storage::delete('public/avatars/' . $user->avatar);
             }
-            $request->avatar->storeAs('avatars',$avatarName, 'public');
+            $request->avatar->storeAs('avatars', $avatarName, 'public');
             $user->update(['avatar' => $avatarName]);
-            Image::make('storage/avatars/'.$user->avatar)->resize(150, 150)->save();
+            Image::make('storage/avatars/' . $user->avatar)->resize(150, 150)->save();
         }
 
         $data = $request->all();
+
         $data['avatar'] = $user->avatar;
+
         $user->update($data);
 
         return redirect()->route('admin.profiles.index');
